tag = rsp-20240830

+++ crates/transaction-pool/src/maintain.rs
@@ -4,22 +4,22 @@ use crate::{
     blobstore::{BlobStoreCanonTracker, BlobStoreUpdates},
     error::PoolError,
     metrics::MaintainPoolMetrics,
-    traits::{CanonicalStateUpdate, ChangedAccount, TransactionPool, TransactionPoolExt},
-    BlockInfo,
+    traits::{CanonicalStateUpdate, TransactionPool, TransactionPoolExt},
+    BlockInfo, PoolTransaction,
 };
 use futures_util::{
     future::{BoxFuture, Fuse, FusedFuture},
     FutureExt, Stream, StreamExt,
 };
+use reth_chain_state::CanonStateNotification;
+use reth_chainspec::{ChainSpec, ChainSpecProvider};
+use reth_execution_types::ChangedAccount;
+use reth_fs_util::FsPathError;
 use reth_primitives::{
-    fs::FsPathError, Address, BlockHash, BlockNumber, BlockNumberOrTag,
-    FromRecoveredPooledTransaction, FromRecoveredTransaction, IntoRecoveredTransaction,
+    Address, BlockHash, BlockNumber, BlockNumberOrTag, IntoRecoveredTransaction,
     PooledTransactionsElementEcRecovered, TransactionSigned,
 };
-use reth_provider::{
-    BlockReaderIdExt, BundleStateWithReceipts, CanonStateNotification, ChainSpecProvider,
-    ProviderError, StateProviderFactory,
-};
+use reth_storage_api::{errors::provider::ProviderError, BlockReaderIdExt, StateProviderFactory};
 use reth_tasks::TaskSpawner;
 use std::{
     borrow::Borrow,
@@ -73,7 +73,12 @@ pub fn maintain_transaction_pool_future<Client, P, St, Tasks>(
     config: MaintainPoolConfig,
 ) -> BoxFuture<'static, ()>
 where
-    Client: StateProviderFactory + BlockReaderIdExt + ChainSpecProvider + Clone + Send + 'static,
+    Client: StateProviderFactory
+        + BlockReaderIdExt
+        + ChainSpecProvider<ChainSpec = ChainSpec>
+        + Clone
+        + Send
+        + 'static,
     P: TransactionPoolExt + 'static,
     St: Stream<Item = CanonStateNotification> + Send + Unpin + 'static,
     Tasks: TaskSpawner + 'static,
@@ -94,7 +99,12 @@ pub async fn maintain_transaction_pool<Client, P, St, Tasks>(
     task_spawner: Tasks,
     config: MaintainPoolConfig,
 ) where
-    Client: StateProviderFactory + BlockReaderIdExt + ChainSpecProvider + Clone + Send + 'static,
+    Client: StateProviderFactory
+        + BlockReaderIdExt
+        + ChainSpecProvider<ChainSpec = ChainSpec>
+        + Clone
+        + Send
+        + 'static,
     P: TransactionPoolExt + 'static,
     St: Stream<Item = CanonStateNotification> + Send + Unpin + 'static,
     Tasks: TaskSpawner + 'static,
@@ -109,7 +119,7 @@ pub async fn maintain_transaction_pool<Client, P, St, Tasks>(
             last_seen_block_hash: latest.hash(),
             last_seen_block_number: latest.number,
             pending_basefee: latest
-                .next_block_base_fee(chain_spec.base_fee_params(latest.timestamp + 12))
+                .next_block_base_fee(chain_spec.base_fee_params_at_timestamp(latest.timestamp + 12))
                 .unwrap_or_default(),
             pending_blob_fee: latest.next_block_blob_fee(),
         };
@@ -186,13 +196,18 @@ pub async fn maintain_transaction_pool<Client, P, St, Tasks>(
         if let Some(finalized) =
             last_finalized_block.update(client.finalized_block_number().ok().flatten())
         {
-            match blob_store_tracker.on_finalized_block(finalized) {
-                BlobStoreUpdates::None => {}
-                BlobStoreUpdates::Finalized(blobs) => {
-                    metrics.inc_deleted_tracked_blobs(blobs.len());
-                    // remove all finalized blobs from the blob store
-                    pool.delete_blobs(blobs);
-                }
+            if let BlobStoreUpdates::Finalized(blobs) =
+                blob_store_tracker.on_finalized_block(finalized)
+            {
+                metrics.inc_deleted_tracked_blobs(blobs.len());
+                // remove all finalized blobs from the blob store
+                pool.delete_blobs(blobs);
+                // and also do periodic cleanup
+                let pool = pool.clone();
+                task_spawner.spawn_blocking(Box::pin(async move {
+                    debug!(target: "txpool", finalized_block = %finalized, "cleaning up blob store");
+                    pool.cleanup_blobs();
+                }));
             }
         }
 
@@ -227,7 +242,7 @@ pub async fn maintain_transaction_pool<Client, P, St, Tasks>(
             Some(Ok(Err(res))) => {
                 // Failed to load accounts from state
                 let (accs, err) = *res;
-                debug!(target: "txpool", ?err, "failed to load accounts");
+                debug!(target: "txpool", %err, "failed to load accounts");
                 dirty_addresses.extend(accs);
             }
             Some(Err(_)) => {
@@ -259,13 +274,15 @@ pub async fn maintain_transaction_pool<Client, P, St, Tasks>(
 
                 // fees for the next block: `new_tip+1`
                 let pending_block_base_fee = new_tip
-                    .next_block_base_fee(chain_spec.base_fee_params(new_tip.timestamp + 12))
+                    .next_block_base_fee(
+                        chain_spec.base_fee_params_at_timestamp(new_tip.timestamp + 12),
+                    )
                     .unwrap_or_default();
                 let pending_block_blob_fee = new_tip.next_block_blob_fee();
 
                 // we know all changed account in the new chain
                 let new_changed_accounts: HashSet<_> =
-                    changed_accounts_iter(new_state).map(ChangedAccountEntry).collect();
+                    new_state.changed_accounts().map(ChangedAccountEntry).collect();
 
                 // find all accounts that were changed in the old chain but _not_ in the new chain
                 let missing_changed_acc = old_state
@@ -286,7 +303,7 @@ pub async fn maintain_transaction_pool<Client, P, St, Tasks>(
                             let (addresses, err) = *err;
                             debug!(
                                 target: "txpool",
-                                ?err,
+                                %err,
                                 "failed to load missing changed accounts at new tip: {:?}",
                                 new_tip.hash()
                             );
@@ -323,13 +340,12 @@ pub async fn maintain_transaction_pool<Client, P, St, Tasks>(
                                     )
                                     .ok()
                                 })
-                                .map(
-                                    <P as TransactionPool>::Transaction::from_recovered_pooled_transaction,
-                                )
+                                .map(|tx| {
+                                    <<P as TransactionPool>::Transaction as PoolTransaction>::from_pooled(tx)
+                                })
                         } else {
-                            Some(<P as TransactionPool>::Transaction::from_recovered_transaction(
-                                tx,
-                            ))
+
+                            <P::Transaction as PoolTransaction>::try_from_consensus(tx).ok()
                         }
                     })
                     .collect::<Vec<_>>();
@@ -341,7 +357,7 @@ pub async fn maintain_transaction_pool<Client, P, St, Tasks>(
                     pending_block_blob_fee,
                     changed_accounts,
                     // all transactions mined in the new chain need to be removed from the pool
-                    mined_transactions: new_mined_transactions.into_iter().collect(),
+                    mined_transactions: new_blocks.transaction_hashes().collect(),
                 };
                 pool.on_canonical_state_change(update);
 
@@ -364,7 +380,9 @@ pub async fn maintain_transaction_pool<Client, P, St, Tasks>(
 
                 // fees for the next block: `tip+1`
                 let pending_block_base_fee = tip
-                    .next_block_base_fee(chain_spec.base_fee_params(tip.timestamp + 12))
+                    .next_block_base_fee(
+                        chain_spec.base_fee_params_at_timestamp(tip.timestamp + 12),
+                    )
                     .unwrap_or_default();
                 let pending_block_blob_fee = tip.next_block_blob_fee();
 
@@ -398,7 +416,7 @@ pub async fn maintain_transaction_pool<Client, P, St, Tasks>(
                 }
 
                 let mut changed_accounts = Vec::with_capacity(state.state().len());
-                for acc in changed_accounts_iter(state) {
+                for acc in state.changed_accounts() {
                     // we can always clear the dirty flag for this account
                     dirty_addresses.remove(&acc.address);
                     changed_accounts.push(acc);
@@ -474,11 +492,11 @@ impl MaintainedPoolState {
     /// Returns `true` if the pool is assumed to be out of sync with the current state.
     #[inline]
     const fn is_drifted(&self) -> bool {
-        matches!(self, MaintainedPoolState::Drifted)
+        matches!(self, Self::Drifted)
     }
 }
 
-/// A unique ChangedAccount identified by its address that can be used for deduplication
+/// A unique `ChangedAccount` identified by its address that can be used for deduplication
 #[derive(Eq)]
 struct ChangedAccountEntry(ChangedAccount);
 
@@ -519,10 +537,11 @@ fn load_accounts<Client, I>(
     addresses: I,
 ) -> Result<LoadedAccounts, Box<(HashSet<Address>, ProviderError)>>
 where
-    I: Iterator<Item = Address>,
+    I: IntoIterator<Item = Address>,
 
     Client: StateProviderFactory,
 {
+    let addresses = addresses.into_iter();
     let mut res = LoadedAccounts::default();
     let state = match client.history_by_block_hash(at) {
         Ok(state) => state,
@@ -542,16 +561,6 @@ where
     Ok(res)
 }
 
-/// Extracts all changed accounts from the BundleState
-fn changed_accounts_iter(
-    state: &BundleStateWithReceipts,
-) -> impl Iterator<Item = ChangedAccount> + '_ {
-    state
-        .accounts_iter()
-        .filter_map(|(addr, acc)| acc.map(|acc| (addr, acc)))
-        .map(|(address, acc)| ChangedAccount { address, nonce: acc.nonce, balance: acc.balance })
-}
-
 /// Loads transactions from a file, decodes them from the RLP format, and inserts them
 /// into the transaction pool on node boot up.
 /// The file is removed after the transactions have been successfully processed.
@@ -567,7 +576,7 @@ where
     }
 
     debug!(target: "txpool", txs_file =?file_path, "Check local persistent storage for saved transactions");
-    let data = reth_primitives::fs::read(file_path)?;
+    let data = reth_fs_util::read(file_path)?;
 
     if data.is_empty() {
         return Ok(())
@@ -577,12 +586,17 @@ where
 
     let pool_transactions = txs_signed
         .into_iter()
-        .filter_map(|tx| tx.try_ecrecovered().map(<P::Transaction>::from_recovered_transaction))
+        .filter_map(|tx| tx.try_ecrecovered())
+        .filter_map(|tx| {
+            // Filter out errors
+            <P::Transaction as PoolTransaction>::try_from_consensus(tx).ok()
+        })
         .collect::<Vec<_>>();
+
     let outcome = pool.add_transactions(crate::TransactionOrigin::Local, pool_transactions).await;
 
     info!(target: "txpool", txs_file =?file_path, num_txs=%outcome.len(), "Successfully reinserted local transactions from file");
-    reth_primitives::fs::remove_file(file_path)?;
+    reth_fs_util::remove_file(file_path)?;
     Ok(())
 }
 
@@ -602,12 +616,12 @@ where
         .collect::<Vec<_>>();
 
     let num_txs = local_transactions.len();
-    let mut buf = alloy_rlp::BytesMut::new();
+    let mut buf = Vec::new();
     alloy_rlp::encode_list(&local_transactions, &mut buf);
     info!(target: "txpool", txs_file =?file_path, num_txs=%num_txs, "Saving current local transactions");
     let parent_dir = file_path.parent().map(std::fs::create_dir_all).transpose();
 
-    match parent_dir.map(|_| reth_primitives::fs::write(file_path, buf)) {
+    match parent_dir.map(|_| reth_fs_util::write(file_path, buf)) {
         Ok(_) => {
             info!(target: "txpool", txs_file=?file_path, "Wrote local transactions to file");
         }
@@ -657,17 +671,16 @@ pub async fn backup_local_transactions_task<P>(
     drop(graceful_guard)
 }
 
-#[cfg(not(feature = "optimism"))]
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::{
         blobstore::InMemoryBlobStore, validate::EthTransactionValidatorBuilder,
-        CoinbaseTipOrdering, EthPooledTransaction, Pool, PoolTransaction, TransactionOrigin,
+        CoinbaseTipOrdering, EthPooledTransaction, Pool, TransactionOrigin,
     };
-    use reth_primitives::{
-        fs, hex, FromRecoveredPooledTransaction, PooledTransactionsElement, MAINNET, U256,
-    };
+    use reth_chainspec::MAINNET;
+    use reth_fs_util as fs;
+    use reth_primitives::{hex, PooledTransactionsElement, U256};
     use reth_provider::test_utils::{ExtendedAccount, MockEthProvider};
     use reth_tasks::TaskManager;
 
@@ -687,11 +700,9 @@ mod tests {
         let temp_dir = tempfile::tempdir().unwrap();
         let transactions_path = temp_dir.path().join(FILENAME).with_extension(EXTENSION);
         let tx_bytes = hex!("02f87201830655c2808505ef61f08482565f94388c818ca8b9251b393131c08a736a67ccb192978801049e39c4b5b1f580c001a01764ace353514e8abdfb92446de356b260e3c1225b73fc4c8876a6258d12a129a04f02294aa61ca7676061cd99f29275491218b4754b46a0248e5e42bc5091f507");
-        let tx = PooledTransactionsElement::decode_enveloped(tx_bytes.into()).unwrap();
+        let tx = PooledTransactionsElement::decode_enveloped(&mut &tx_bytes[..]).unwrap();
         let provider = MockEthProvider::default();
-        let transaction = EthPooledTransaction::from_recovered_pooled_transaction(
-            tx.try_into_ecrecovered().unwrap(),
-        );
+        let transaction: EthPooledTransaction = tx.try_into_ecrecovered().unwrap().into();
         let tx_to_cmp = transaction.clone();
         let sender = hex!("1f9090aaE28b8a3dCeaDf281B0F12828e676c326").into();
         provider.add_account(sender, ExtendedAccount::new(42, U256::MAX));
@@ -730,5 +741,47 @@ mod tests {
         assert_eq!(txs.len(), 1);
 
         temp_dir.close().unwrap();
+    }
+
+    #[test]
+    fn test_update_with_higher_finalized_block() {
+        let mut tracker = FinalizedBlockTracker::new(Some(10));
+        assert_eq!(tracker.update(Some(15)), Some(15));
+        assert_eq!(tracker.last_finalized_block, Some(15));
+    }
+
+    #[test]
+    fn test_update_with_lower_finalized_block() {
+        let mut tracker = FinalizedBlockTracker::new(Some(20));
+        assert_eq!(tracker.update(Some(15)), None);
+        assert_eq!(tracker.last_finalized_block, Some(15));
+    }
+
+    #[test]
+    fn test_update_with_equal_finalized_block() {
+        let mut tracker = FinalizedBlockTracker::new(Some(20));
+        assert_eq!(tracker.update(Some(20)), None);
+        assert_eq!(tracker.last_finalized_block, Some(20));
+    }
+
+    #[test]
+    fn test_update_with_no_last_finalized_block() {
+        let mut tracker = FinalizedBlockTracker::new(None);
+        assert_eq!(tracker.update(Some(10)), Some(10));
+        assert_eq!(tracker.last_finalized_block, Some(10));
+    }
+
+    #[test]
+    fn test_update_with_no_new_finalized_block() {
+        let mut tracker = FinalizedBlockTracker::new(Some(10));
+        assert_eq!(tracker.update(None), None);
+        assert_eq!(tracker.last_finalized_block, Some(10));
+    }
+
+    #[test]
+    fn test_update_with_no_finalized_blocks() {
+        let mut tracker = FinalizedBlockTracker::new(None);
+        assert_eq!(tracker.update(None), None);
+        assert_eq!(tracker.last_finalized_block, None);
     }
 }
